@@ -12,20 +12,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "~13.3.2",
-    "@angular/animations": "^13.3.2",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/animations": "^14.2.12",
     "@angular/cdk": "^13.3.2",
-    "@angular/common": "^13.3.2",
-    "@angular/compiler": "^13.3.2",
-    "@angular/core": "^13.3.2",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
     "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "^13.3.2",
-    "@angular/localize": "^13.3.2",
+    "@angular/forms": "^14.2.12",
+    "@angular/localize": "^14.2.12",
     "@angular/material": "^13.3.2",
-    "@angular/platform-browser": "^13.3.2",
-    "@angular/platform-browser-dynamic": "^13.3.2",
-    "@angular/router": "^13.3.2",
-    "@angular/service-worker": "^13.3.2",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
+    "@angular/service-worker": "^14.2.12",
     "@apollo/client": "^3.0.0",
     "@auth0/angular-jwt": "^3.0.1",
     "@ng-bootstrap/ng-bootstrap": "^12.0.1",
@@ -56,9 +56,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular/cli": "^13.3.2",
-    "@angular/compiler-cli": "^13.3.2",
-    "@angular/language-service": "^13.3.2",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
+    "@angular/language-service": "^14.2.12",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "^12.11.1",
