@@ -43,7 +43,7 @@ import { QueryLanguageComponent } from './query-language/query-language.componen
 import { TrackhubsSubmissionComponent } from './trackhubs-submission/trackhubs-submission.component';
 import { NextflowSubmissionComponent } from './nextflow-submission/nextflow-submission.component';
 import { LocalGenomeBrowserComponent } from './local-genome-browser/local-genome-browser.component';
-// import {GraphqlComponent} from './graphql/graphql.component';
+import {GraphqlComponent} from './graphql/graphql.component';
 
 const routes: Routes = [
   {path: '', redirectTo: 'home', pathMatch: 'full'},
@@ -90,7 +90,7 @@ const routes: Routes = [
   {path: 'trackhubs', component: TrackhubsSubmissionComponent},
   {path: 'nextflowSubmission', component: NextflowSubmissionComponent},
   {path: 'genome_browser', component: LocalGenomeBrowserComponent},
-  // {path: 'graphql', component: GraphqlComponent},
+  {path: 'graphql', component: GraphqlComponent},
   {path: '404', component: NonExistingComponent},
   {path: '**', component: NonExistingComponent}
 ];
